refactor(RI-8284): move AddKeyArray Props into AddKeyArray.types.ts
Follow the component file-structure convention (types live in a sibling
.types.ts) instead of declaring the props interface inline.
