+++ redisinsight/ui/src/pages/browser/components/add-key/AddKeyArray/index.ts
@@ -1,4 +1,4 @@
 import AddKeyArray from './AddKeyArray'
 
 export default AddKeyArray
-export type { Props } from './AddKeyArray'
+export type { Props } from './AddKeyArray.types'
